Upgraded minecraft to 1.20.2 	modified:   gradle.properties 	modified:   gradlew 	modified:   src/main/java/net/lionarius/skinrestorer/SkinRestorer.java 	modified:   src/main/java/net/lionarius/skinrestorer/mixin/PlayerManagerMixin.java 	deleted:    src/main/java/net/lionarius/skinrestorer/mixin/ServerLoginNetworkHandlerMixin.java 	modified:   src/main/resources/skinrestorer.mixins.json
