+++ gradle.properties
@@ -2,10 +2,10 @@
 org.gradle.jvmargs=-Xmx1G
 # Fabric Properties
 # check these on https://fabricmc.net/versions.html
-minecraft_version=1.20
-yarn_mappings=1.20+build.1
-loader_version=0.14.21
+minecraft_version=1.20.2
+yarn_mappings=1.20.2+build.1
+loader_version=0.14.22
 # Mod Properties
-mod_version=1.2.3
+mod_version=1.2.4
 maven_group=net.lionarius
 archives_base_name=skin-restorer
